requirements.txt
prometheus_client
litellm>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
